minimum size sub array

+++ src/main/java/Searching/Questions/MinimumSizeSubArray.java
@@ -7,22 +7,22 @@ public class MinimumSizeSubArray {
         //https://dashboard.programmingpathshala.com/practice/question?questionId=121&sectionId=1&moduleId=2&topicId=6&subtopicId=82&homeworkId=17
 
         Scanner s = new Scanner(System.in);
-        int n = s.nextInt();
-        int[] arr = new int[n];
+        long n = s.nextLong();
+        long[] arr = new long[(int) n];
         for (int i=0; i<n; i++){
-            arr[i] = s.nextInt();
+            arr[i] = s.nextLong();
         }
-        int T = s.nextInt();
+        long T = s.nextLong();
         while (T-- > 0){
-            int target = s.nextInt();
+            long target = s.nextLong();
             System.out.println(findMinimumSizeSubArray(arr, n, target));
         }
     }
 
-    private static int findMinimumSizeSubArray(int[] arr, int n, int target) {
-        int l=0, h=n;
+    private static long findMinimumSizeSubArray(long[] arr, long n, long target) {
+        long l=0, h=n;
         while (l<=h){
-            int mid = (l+h)/2;
+            long mid = (l+h)/2;
             boolean r = isPossible(arr, n, target, mid);
             if (!r){
                 l = mid+1;
@@ -35,10 +35,10 @@ public class MinimumSizeSubArray {
                 }
             }
         }
-        return 0;
+        return 0L;
     }
 
-    private static boolean isPossible(int[] arr, int n, int target, int mid) {
+    private static boolean isPossible(long[] arr, long n, long target, long mid) {
         if (mid == n+1){
             return false;
         }
@@ -47,7 +47,7 @@ public class MinimumSizeSubArray {
             sum += arr[i];
         }
         if (sum >= target) return true;
-        int s=0, e=mid-1;
+        int s=0, e= (int) (mid-1);
         while (e<n-1){
             sum -= arr[s];
             s++;
